Dualstack Registry
Rather than have two images for Registry v1 and Registry v2, this
installs everything in the same image, and the only difference between
the two registries is which one will accept request on the server's
default hostname.

With some work on Sweetness and API, this should allow running both
registries in parallel, by retaining some metadata in API about which
registry we should ask for a given image. This will be done by accessing
the registry with different hostnames:

- `registry-i-foo.aptible.in`: default registry (depends on the image).
- `dualstack-v1-registry-i-foo.aptible.in`: force v1.
- `dualstack-v2-registry-i-foo.aptible.in`: force v2.