Test connecting via dualstack host
This adds tests that check we can connect to a dualstack registry via
the dualstack endpoints, and that the default endpoint talks to the
correct registry versions.

This also enforces that `SERVER_NAME` be set (and no longer derives it
from the certificate), since we'll need it to know what host a given
request is for.