Support us-east-2
Neither boto nor registry support us-east-2 as a region for S3 in the
versions we're currently using. Upgrading Boto isn't super important
since we're unlikely to be using v1 registries in us-east-2 (but it's
important that the registry at least boots!), but of course upgrading
the v2 registry matters here.

For Boto, I've upgraded Boto itself (and pinned our other dependencies,
just to be safe), and removed the patch that we used to carry from
registry, because that patch is now part of Boto itself AFAICT.

For v2 registry, I've upgraded us to 2.6.1. There don't appear to be any
relevant changes except for us-east-2 support according to the
changelog.